pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PythonBinder"
version = "1.1.0"
description = "Advanced Python to EXE compiler - MERO"
readme = "README.md"
requires-python = ">=3.6"
license = "MIT"
authors = [
    {name = "MERO", email = "qp4rm@telegram.contact"}
]
keywords = ["compiler", "exe", "windows", "pyinstaller", "obfuscation", "MERO"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Compilers"
]

dependencies = [
    "tqdm>=4.50.0",
    "Pillow>=8.0.0"
]

[project.urls]
Homepage = "https://t.me/QP4RM"
Repository = "https://github.com/6x-u/PythonBinder"

[project.scripts]
py-PythonBinder = "pythonbinder.cli:main"

[tool.setuptools]
packages = ["pythonbinder"]
include-package-data = true

[tool.setuptools.package-data]
pythonbinder = ["assets/*", "*.png"]
